feat: allow multiple video formats
