+++ Assets/Scripts/Recording/Config/ConfigReader.cs.meta
